@@ -16,14 +16,14 @@
   },
   "devDependencies": {
     "@types/node": "18.15.13",
-    "@types/react": "^17.0.37",
-    "@types/react-dom": "^17.0.11",
-    "eslint": "^7.32.0",
-    "tailwind-config": "workspace:*",
-    "eslint-config-next": "13.3.0",
-    "tailwindcss": "3.3.1",
-    "postcss": "8.4.23",
+    "@types/react": "^17.0.58",
+    "@types/react-dom": "^17.0.20",
     "autoprefixer": "10.4.14",
-    "typescript": "^4.5.2"
+    "eslint": "^7.32.0",
+    "eslint-config-next": "13.3.0",
+    "postcss": "8.4.23",
+    "tailwind-config": "workspace:*",
+    "tailwindcss": "3.3.1",
+    "typescript": "^4.9.5"
   }
 }
